refactor(gen): move gen to itergen, some util funcs to typeutil package
