+++ telegram/query/channels/participants/query.go
@@ -1,3 +1,3 @@
 package participants
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -result=ChannelsChannelParticipantsClass -package=participants -prefix=Channels -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -result=ChannelsChannelParticipantsClass -package=participants -prefix=Channels -out=queries.gen.go
