+++ telegram/query/contacts/blocked/query.go
@@ -1,3 +1,3 @@
 package blocked
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -result=ContactsBlockedClass -package=blocked -prefix=Contacts -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -result=ContactsBlockedClass -package=blocked -prefix=Contacts -out=queries.gen.go
