+++ telegram/query/dialogs/query.go
@@ -1,3 +1,3 @@
 package dialogs
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -result=MessagesDialogsClass -package=dialogs -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -result=MessagesDialogsClass -package=dialogs -out=queries.gen.go
