+++ telegram/query/internal/itergen/_template/gen.tmpl
