+++ telegram/query/internal/itergen/collect.go
@@ -9,6 +9,8 @@ import (
 
 	"golang.org/x/tools/go/packages"
 	"golang.org/x/xerrors"
+
+	"github.com/gotd/td/telegram/query/internal/typeutil"
 )
 
 type method struct {
@@ -28,7 +30,8 @@ type collector struct {
 	requiredByIter     []string
 	required           map[string]string
 
-	implsCache map[string][]*types.Named
+	pkg    *packages.Package
+	ifaces *typeutil.Interfaces
 
 	iface          *types.Interface
 	resultTypeName string
@@ -52,7 +55,7 @@ func (c *collectorConfig) fromFlags(set *flag.FlagSet) {
 	set.StringVar(&c.PkgName, "package", "messages", "name of package name to generate")
 }
 
-func newCollector(cfg collectorConfig) *collector {
+func newCollector(pkg *packages.Package, cfg collectorConfig) *collector {
 	intGetter := types.NewSignature(nil, nil,
 		types.NewTuple(types.NewVar(0, nil, "", types.Typ[types.Int])), false) // func() int
 	methods := []*types.Func{
@@ -96,7 +99,8 @@ func newCollector(cfg collectorConfig) *collector {
 		canFillFromRequest: canFillFromRequest,
 		requiredByIter:     requiredByIter,
 		required:           required,
-		implsCache:         map[string][]*types.Named{},
+		ifaces:             typeutil.NewInterfaces(pkg),
+		pkg:                pkg,
 		iface:              match,
 		resultTypeName:     cfg.ResultName,
 		elemName:           cfg.ElemName,
@@ -105,10 +109,10 @@ func newCollector(cfg collectorConfig) *collector {
 	}
 }
 
-func (c *collector) methods(pkg *packages.Package) ([]method, error) { // nolint:gocognit
+func (c *collector) methods() ([]method, error) { // nolint:gocognit
 	var result []method
 
-	for _, def := range pkg.TypesInfo.Defs {
+	for _, def := range c.pkg.TypesInfo.Defs {
 		if def == nil {
 			continue
 		}
@@ -205,16 +209,8 @@ func (c *collector) methods(pkg *packages.Package) ([]method, error) { // nolint
 	return result, nil
 }
 
-func sortParams(p []Param) []Param {
-	sort.SliceStable(p, func(i, j int) bool {
-		return p[i].Name < p[j].Name
-	})
-
-	return p
-}
-
-func (c *collector) Config(pkg *packages.Package) (Config, error) {
-	methods, err := c.collect(pkg)
+func (c *collector) Config() (Config, error) {
+	methods, err := c.collect()
 	if err != nil {
 		return Config{}, xerrors.Errorf("collect: %w", err)
 	}
@@ -227,8 +223,8 @@ func (c *collector) Config(pkg *packages.Package) (Config, error) {
 	}, nil
 }
 
-func (c *collector) collect(pkg *packages.Package) ([]Method, error) {
-	methods, err := c.methods(pkg)
+func (c *collector) collect() ([]Method, error) {
+	methods, err := c.methods()
 	if err != nil {
 		return nil, xerrors.Errorf("collect types: %w", err)
 	}
@@ -243,8 +239,8 @@ func (c *collector) collect(pkg *packages.Package) ([]Method, error) {
 		m := Method{
 			Name:              method.name,
 			OriginalName:      method.f.Name(),
-			RequestName:       printType(method.reqType),
-			ResultName:        printType(method.resultType),
+			RequestName:       typeutil.PrintType(method.reqType),
+			ResultName:        typeutil.PrintType(method.resultType),
 			AdditionalMapping: mapping,
 			AdditionalParams:  sortParams(method.params),
 			IteratorName:      "Iterator",
@@ -258,7 +254,7 @@ func (c *collector) collect(pkg *packages.Package) ([]Method, error) {
 		}
 		m.RequiredParams = sortParams(m.RequiredParams)
 
-		cases, err := c.collectSpecial(pkg, m)
+		cases, err := c.collectSpecial(m)
 		if err != nil {
 			return nil, xerrors.Errorf("collect special: %w", err)
 		}
