+++ telegram/query/internal/itergen/gen_test.go
@@ -8,12 +8,12 @@ import (
 
 func TestGenerate(t *testing.T) {
 	var out bytes.Buffer
-	if err := generate(context.Background(), &out, newCollector(collectorConfig{
+	if err := generate(context.Background(), &out, collectorConfig{
 		ResultName: "MessagesMessagesClass",
 		ElemName:   "Elem",
 		Prefix:     "Messages",
 		PkgName:    "messages",
-	})); err != nil {
+	}); err != nil {
 		t.Fatal(err)
 	}
 }
