+++ telegram/query/internal/typeutil/doc.go
@@ -1,0 +1,2 @@
+// Package typeutil is a helper package to work with types during codegeneration.
+package typeutil
