+++ telegram/query/messages/query.go
@@ -1,3 +1,3 @@
 package messages
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -out=queries.gen.go
