+++ telegram/query/messages/stickers/featured/query.go
@@ -1,3 +1,3 @@
 package featured
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -result=MessagesFeaturedStickersClass -package=featured -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -result=MessagesFeaturedStickersClass -package=featured -out=queries.gen.go
