+++ telegram/query/photos/query.go
@@ -1,3 +1,3 @@
 package photos
 
-//go:generate go run github.com/gotd/td/telegram/query/internal/gen -result=PhotosPhotosClass -package=photos -prefix=Photos -out=queries.gen.go
+//go:generate go run github.com/gotd/td/telegram/query/internal/itergen -result=PhotosPhotosClass -package=photos -prefix=Photos -out=queries.gen.go
